[Agent] factor operation handler registration
